# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RemnaRay API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/admin/v1/admins": {
      "get": {
        "operationId": "listAdmins",
        "summary": "Administrators (FR-143)",
        "responses": {
          "200": {
            "description": "Administrators"
          },
          "403": {
            "description": "FORBIDDEN"
          }
        }
      },
      "post": {
        "operationId": "createAdmin",
        "responses": {
          "201": {
            "description": "Created"
          },
          "409": {
            "description": "CONFLICT"
          },
          "403": {
            "description": "FORBIDDEN"
          }
        }
      }
    },
    "/api/admin/v1/admins/{id}": {
      "patch": {
        "operationId": "updateAdmin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated"
          },
          "404": {
            "description": "NOT_FOUND"
          },
          "409": {
            "description": "LAST_ADMIN"
          }
        }
      }
    },
    "/api/admin/v1/admins/{id}/deactivate": {
      "post": {
        "operationId": "deactivateAdmin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deactivated"
          },
          "404": {
            "description": "NOT_FOUND"
          },
          "409": {
            "description": "LAST_ADMIN"
          }
        }
      }
    },
    "/api/admin/v1/admins/{id}/reset-password": {
      "post": {
        "operationId": "resetAdminPassword",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password reset"
          },
          "404": {
            "description": "NOT_FOUND"
          }
        }
      }
    },
    "/api/admin/v1/admins/{id}/reset-totp": {
      "post": {
        "operationId": "resetAdminTotp",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "TOTP enrolment cleared"
          },
          "404": {
            "description": "NOT_FOUND"
          }
        }
      }
    },
    "/api/admin/v1/auth/login": {
      "post": {
        "operationId": "adminLogin",
        "responses": {
          "200": {
            "description": "Password challenge"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "423": {
            "description": "Admin locked"
          }
        }
      }
    },
    "/api/admin/v1/auth/logout": {
      "post": {
        "operationId": "adminLogout",
        "responses": {
          "204": {
            "description": "Logged out"
          }
        }
      }
    },
    "/api/admin/v1/auth/me": {
      "get": {
        "operationId": "adminMe",
        "responses": {
          "200": {
            "description": "Current admin and CSRF token"
          },
          "401": {
            "description": "Unauthenticated"
          }
        }
      }
    },
    "/api/admin/v1/auth/totp": {
      "post": {
        "operationId": "adminTotpLogin",
        "responses": {
          "200": {
            "description": "Authenticated admin session"
          },
          "401": {
            "description": "Invalid TOTP"
          }
        }
      }
    },
    "/api/admin/v1/auth/totp/confirm": {
      "post": {
        "operationId": "adminTotpConfirm",
        "responses": {
          "200": {
            "description": "Authenticated first admin session"
          },
          "401": {
            "description": "Invalid TOTP"
          }
        }
      }
    },
    "/api/admin/v1/auth/totp/setup": {
      "post": {
        "operationId": "adminTotpSetup",
        "responses": {
          "200": {
            "description": "TOTP setup QR"
          },
          "401": {
            "description": "Invalid challenge"
          }
        }
      }
    },
    "/api/admin/v1/plans": {
      "get": {
        "operationId": "listPlans",
        "responses": {
          "200": {
            "description": "Plans"
          }
        }
      },
      "post": {
        "operationId": "createPlan",
        "responses": {
          "201": {
            "description": "Created plan"
          }
        }
      }
    },
    "/api/admin/v1/plans/{id}": {
      "patch": {
        "operationId": "updatePlan",
        "responses": {
          "200": {
            "description": "Updated plan"
          }
        }
      },
      "delete": {
        "operationId": "deletePlan",
        "responses": {
          "200": {
            "description": "Deleted plan"
          }
        }
      }
    },
    "/api/admin/v1/settings": {
      "get": {
        "operationId": "getSettings",
        "responses": {
          "200": {
            "description": "Grouped settings"
          }
        }
      },
      "put": {
        "operationId": "updateSettings",
        "responses": {
          "200": {
            "description": "Updated settings"
          }
        }
      }
    },
    "/api/admin/v1/settings/export": {
      "get": {
        "operationId": "exportSettings",
        "responses": {
          "200": {
            "description": "Safe settings snapshot"
          }
        }
      }
    },
    "/api/admin/v1/settings/import": {
      "post": {
        "operationId": "importSettings",
        "responses": {
          "200": {
            "description": "Imported settings"
          }
        }
      }
    },
    "/api/admin/v1/settings/schema": {
      "get": {
        "operationId": "getSettingsSchema",
        "responses": {
          "200": {
            "description": "Settings schema"
          }
        }
      }
    },
    "/api/admin/v1/themes": {
      "get": {
        "operationId": "listThemes",
        "summary": "Themes available in the mounted themes directory",
        "responses": {
          "200": {
            "description": "Theme list"
          },
          "401": {
            "description": "UNAUTHENTICATED"
          }
        }
      }
    },
    "/api/internal/v1/auth/issue-token": {
      "post": {
        "operationId": "issueBotToken",
        "responses": {
          "200": {
            "description": "Bot JWT"
          }
        }
      }
    },
    "/api/internal/v1/ledger/users/{userId}/available": {
      "get": {
        "operationId": "getLedgerAvailable",
        "responses": {
          "200": {
            "description": "Spendable balance in minor units"
          }
        }
      }
    },
    "/api/internal/v1/me/plans": {
      "get": {
        "operationId": "listInternalPlans",
        "summary": "PlanPublic list for the bot",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/internal/v1/payments/expire": {
      "post": {
        "operationId": "expireInvoices",
        "responses": {
          "200": {
            "description": "Expired count"
          }
        }
      }
    },
    "/api/internal/v1/payments/poll-pending": {
      "post": {
        "operationId": "pollPendingInvoices",
        "responses": {
          "200": {
            "description": "Polled count"
          }
        }
      }
    },
    "/api/internal/v1/remnawave/health": {
      "get": {
        "operationId": "remnawaveHealth",
        "responses": {
          "200": {
            "description": "Panel health"
          }
        }
      }
    },
    "/api/internal/v1/remnawave/reconcile": {
      "post": {
        "operationId": "remnawaveReconcile",
        "responses": {
          "200": {
            "description": "Reconcile result"
          }
        }
      }
    },
    "/api/internal/v1/subscriptions/expire": {
      "post": {
        "operationId": "expireSubscriptions",
        "responses": {
          "201": {
            "description": "Expired subscription count"
          }
        }
      }
    },
    "/api/internal/v1/support/forward": {
      "post": {
        "operationId": "forwardSupportMessage",
        "summary": "Forward a support message to the configured chat",
        "responses": {
          "204": {
            "description": "Forwarded"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/internal/v1/users/upsert": {
      "post": {
        "operationId": "upsertInternalUser",
        "responses": {
          "200": {
            "description": "Upserted user"
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout",
        "responses": {
          "204": {
            "description": "Logged out"
          }
        }
      }
    },
    "/api/v1/auth/telegram": {
      "post": {
        "operationId": "telegramLogin",
        "responses": {
          "200": {
            "description": "Authenticated user"
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "operationId": "getMe",
        "summary": "UserMe",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "UNAUTHENTICATED"
          }
        }
      },
      "patch": {
        "operationId": "patchMe",
        "summary": "Update language, email and marketing opt-out",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "UNAUTHENTICATED"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/v1/me/anonymize-request": {
      "post": {
        "operationId": "requestMyAnonymization",
        "summary": "Section 19.5 self-request; records an audit entry for an administrator",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "UNAUTHENTICATED"
          }
        }
      }
    },
    "/api/v1/me/invoices": {
      "post": {
        "operationId": "createMyInvoice",
        "summary": "Create an invoice",
        "responses": {
          "201": {
            "description": "Invoice"
          },
          "409": {
            "description": "PLAN_UNAVAILABLE | PROVIDER_UNAVAILABLE | INSUFFICIENT_FUNDS | PROMO_*"
          },
          "400": {
            "description": "TOPUP_AMOUNT_OUT_OF_RANGE"
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/v1/me/invoices/{id}": {
      "get": {
        "operationId": "getMyInvoice",
        "summary": "Invoice",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT_FOUND"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/v1/me/invoices/{id}/cancel": {
      "post": {
        "operationId": "cancelMyInvoice",
        "summary": "Cancel a pending invoice",
        "responses": {
          "200": {
            "description": "OK"
          },
          "409": {
            "description": "INVOICE_NOT_PENDING"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/v1/me/invoices/{id}/check": {
      "post": {
        "operationId": "checkMyInvoice",
        "summary": "Re-fetch the provider status",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "NOT_FOUND"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/v1/me/payment-methods": {
      "get": {
        "operationId": "listMyPaymentMethods",
        "summary": "Providers offered to this user",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "UNAUTHENTICATED"
          }
        }
      }
    },
    "/api/v1/me/plan-change/quote": {
      "get": {
        "operationId": "getMyPlanChangeQuote",
        "summary": "Credit and amount to pay",
        "responses": {
          "200": {
            "description": "OK"
          },
          "409": {
            "description": "PLAN_CHANGE_NOT_ALLOWED"
          }
        },
        "parameters": [
          {
            "name": "planId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/v1/me/promocodes/preview": {
      "post": {
        "operationId": "previewMyPromocode",
        "summary": "Discount preview for a plan",
        "responses": {
          "200": {
            "description": "OK"
          },
          "409": {
            "description": "PROMO_*"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/v1/me/promocodes/redeem": {
      "post": {
        "operationId": "redeemMyPromocode",
        "summary": "Reserve a promo code for the next purchase",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "PROMO_NOT_FOUND"
          },
          "409": {
            "description": "PROMO_*"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/v1/me/referrals": {
      "get": {
        "operationId": "getMyReferrals",
        "summary": "Referral summary and program terms",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "UNAUTHENTICATED"
          }
        }
      }
    },
    "/api/v1/me/referrals/list": {
      "get": {
        "operationId": "listMyReferrals",
        "summary": "Invited users with masked names",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "UNAUTHENTICATED"
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/me/subscription": {
      "get": {
        "operationId": "getMySubscription",
        "summary": "Subscription, panel snapshot and client links",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "UNAUTHENTICATED"
          }
        }
      }
    },
    "/api/v1/me/subscription/devices": {
      "get": {
        "operationId": "listMyDevices",
        "summary": "HWID devices from the panel",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "UNAUTHENTICATED"
          }
        }
      }
    },
    "/api/v1/me/subscription/devices/{hwid}": {
      "delete": {
        "operationId": "removeMyDevice",
        "summary": "Remove one HWID device",
        "responses": {
          "204": {
            "description": "Removed"
          },
          "403": {
            "description": "FORBIDDEN"
          }
        },
        "parameters": [
          {
            "name": "hwid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/me/subscription/qr": {
      "get": {
        "operationId": "getMySubscriptionQr",
        "summary": "512x512 PNG of the subscription link",
        "responses": {
          "200": {
            "description": "PNG",
            "content": {
              "image/png": {}
            }
          },
          "404": {
            "description": "NOT_FOUND"
          }
        }
      }
    },
    "/api/v1/me/subscription/revoke": {
      "post": {
        "operationId": "revokeMySubscription",
        "summary": "Reset the subscription link",
        "responses": {
          "200": {
            "description": "OK"
          },
          "429": {
            "description": "REVOKE_RATE_LIMITED"
          }
        }
      }
    },
    "/api/v1/me/topup-config": {
      "get": {
        "operationId": "getTopupConfig",
        "summary": "Presets and allowed range",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "UNAUTHENTICATED"
          }
        }
      }
    },
    "/api/v1/me/transactions": {
      "get": {
        "operationId": "listMyTransactions",
        "summary": "Cursor-paginated transactions",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "UNAUTHENTICATED"
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/me/trial": {
      "post": {
        "operationId": "startMyTrial",
        "summary": "One-time trial",
        "responses": {
          "200": {
            "description": "OK"
          },
          "409": {
            "description": "TRIAL_ALREADY_USED | TRIAL_NOT_ELIGIBLE"
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/v1/public/config": {
      "get": {
        "operationId": "getPublicConfig",
        "summary": "Brand, locales, features and legal revision dates",
        "responses": {
          "200": {
            "description": "Public configuration"
          }
        }
      }
    },
    "/api/v1/public/i18n/{lang}/{namespace}": {
      "get": {
        "operationId": "getPublicMessages",
        "summary": "Merged messages for one namespace (overrides applied)",
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["ru", "en"]
            }
          },
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Messages",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "Not modified"
          },
          "404": {
            "description": "NOT_FOUND"
          }
        }
      }
    },
    "/api/v1/public/legal/{doc}": {
      "get": {
        "operationId": "getPublicLegal",
        "summary": "Legal document markdown with brand substitutions",
        "parameters": [
          {
            "name": "doc",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["terms", "privacy", "offer"]
            }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["ru", "en"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Legal document"
          },
          "404": {
            "description": "NOT_FOUND"
          }
        }
      }
    },
    "/api/v1/public/plans": {
      "get": {
        "operationId": "getPublicPlans",
        "responses": {
          "200": {
            "description": "Public plans"
          }
        }
      }
    },
    "/api/v1/public/theme": {
      "get": {
        "operationId": "getPublicTheme",
        "summary": "Active theme tokens and asset URLs (section 18.2)",
        "parameters": [
          {
            "name": "If-None-Match",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Theme tokens",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "Not modified"
          }
        }
      }
    },
    "/api/v1/r/{code}": {
      "get": {
        "operationId": "followReferralLink",
        "summary": "Stores the referral code for 30 days and redirects to the site",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to /",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/webhooks/remnawave": {
      "post": {
        "operationId": "remnawaveWebhook",
        "responses": {
          "200": {
            "description": "Webhook result"
          }
        }
      }
    },
    "/webhooks/{provider}": {
      "post": {
        "operationId": "paymentWebhook",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Acknowledged"
          },
          "400": {
            "description": "Invalid signature"
          }
        }
      }
    }
  }
}